change endpoint location to charts
